+++ src/inspect_ai/_eval/score.py
@@ -286,6 +286,7 @@ async def score_async(
             resolved_scorers,
             log_metrics,
             scorer_names,
+            log.results.early_stopping if log.results else None,
         )
 
         # Since the metrics calculation above is only be done using the scorers