+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- [Early Stopping](https://inspect.aisi.org.uk/early-stopping.html.md) API for ending tasks early based on previously scored samples.
 - Logging: Support for writing logs to Azure Blob Storage (`az://`).
 - React Agent: Remove newlines from default prompts.
 - Bugfix: Copy `metadata` field to new eval for `eval-retry`.